requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-bootstrap5==24.3
django-braces==1.16.0
django-crispy-forms==2.3
lxml==5.3.0
python-docx==1.1.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
